requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.7.0
python-multipart>=0.0.9
anthropic>=0.28.0
scikit-learn>=1.4.0
python-docx>=1.1.0

# Optional — comment out if not using
# gpt-researcher>=0.10.0
# supabase>=2.4.0
# notion-client>=2.2.1
# google-generativeai>=0.7.0
# openai>=1.30.0
# weasyprint>=62.0
